refactor(app): split explain reducers
