+++ src/app/update/explain/mod.rs
@@ -1,307 +1,50 @@
+mod analyze;
+mod helpers;
+mod request;
+mod result;
+mod scroll;
+mod tabs;
+
 use std::time::Instant;
 
-use crate::cmd::effect::Effect;
 use crate::model::app_state::AppState;
-use crate::model::explain_context::ExplainContext;
-use crate::model::shared::text_input::TextInputLike;
-use crate::model::sql_editor::modal::{SqlModalStatus, SqlModalTab};
-use crate::policy::sql::statement_classifier::{self, StatementKind};
-use crate::policy::write::sql_risk::{ConfirmationType, evaluate_sql_risk, split_statements};
 use crate::services::AppServices;
-use crate::update::action::{Action, ScrollAmount, ScrollTarget};
+use crate::update::action::Action;
 use crate::update::dispatch_result::DispatchResult;
 
-fn is_multi_statement(content: &str) -> bool {
-    split_statements(content).len() > 1
-}
+#[cfg(test)]
+use crate::cmd::effect::Effect;
+#[cfg(test)]
+use crate::model::sql_editor::modal::{SqlModalStatus, SqlModalTab};
+#[cfg(test)]
+use crate::update::action::{ScrollAmount, ScrollTarget};
 
-fn mark_explain_unavailable(state: &mut AppState, services: &AppServices) {
-    state
-        .explain
-        .set_error("EXPLAIN is unavailable for this database".to_string());
-    let tab = services
-        .db_capabilities
-        .normalize_sql_modal_tab(state.sql_modal.active_tab());
-    state.sql_modal.set_active_tab(tab);
-}
-
-fn show_explain_error_on_plan(state: &mut AppState, message: impl Into<String>) {
-    state.explain.set_error(message.into());
-    state.sql_modal.set_active_tab(SqlModalTab::Plan);
-}
-
-fn begin_explain_running(state: &mut AppState, now: Instant) {
-    state.sql_modal.begin_adhoc_running();
-    state.sql_modal.set_active_tab(SqlModalTab::Plan);
-    state.explain.reset();
-    state.query.begin_running(now);
-}
-
-fn finish_explain_success(
+pub fn dispatch_explain(
     state: &mut AppState,
-    plan_text: String,
-    is_analyze: bool,
-    execution_time_ms: u64,
-    query: &str,
-) {
-    state
-        .explain
-        .set_plan(plan_text, is_analyze, execution_time_ms, query);
-    state.sql_modal.enter_normal();
-    state.sql_modal.set_active_tab(SqlModalTab::Plan);
-    state.query.mark_idle();
+    action: &Action,
+    now: Instant,
+    services: &AppServices,
+) -> DispatchResult {
+    request::reduce_request(state, action, now, services)
+        .or_else(|| analyze::reduce_analyze(state, action, now, services))
+        .or_else(|| result::reduce_result(state, action, now))
+        .or_else(|| scroll::reduce_scroll(state, action, now))
+        .or_else(|| tabs::reduce_tabs(state, action, now, services))
 }
 
-fn finish_explain_error(state: &mut AppState, error: impl Into<String>) {
-    state.explain.set_error(error.into());
-    state.sql_modal.enter_normal();
-    state.sql_modal.set_active_tab(SqlModalTab::Plan);
-    state.query.mark_idle();
-}
-
-fn reject_unsupported_explain(state: &mut AppState, services: &AppServices) -> bool {
-    if services.db_capabilities.supports_explain() {
-        return false;
-    }
-
-    mark_explain_unavailable(state, services);
-    true
-}
-
+#[cfg(test)]
 pub fn reduce_explain_with_services(
     state: &mut AppState,
     action: &Action,
     now: Instant,
     services: &AppServices,
 ) -> DispatchResult {
-    match action {
-        Action::ExplainRequest => {
-            if reject_unsupported_explain(state, services) {
-                return DispatchResult::handled();
-            }
-            let content = state.sql_modal.editor.content().trim().to_string();
-            if content.is_empty() {
-                return DispatchResult::handled();
-            }
-            let Some(dsn) = state.session.dsn.clone() else {
-                return DispatchResult::handled();
-            };
-            if matches!(state.sql_modal.status(), SqlModalStatus::Running) {
-                return DispatchResult::handled();
-            }
-            if is_multi_statement(&content) {
-                show_explain_error_on_plan(state, "EXPLAIN does not support multiple statements");
-                return DispatchResult::handled();
-            }
-
-            let Some(query) = services.sql_dialect.build_explain_sql(&content) else {
-                mark_explain_unavailable(state, services);
-                return DispatchResult::handled();
-            };
-            begin_explain_running(state, now);
-
-            DispatchResult::handled_with(vec![Effect::ExecuteExplain {
-                dsn,
-                query,
-                is_analyze: false,
-                read_only: true,
-            }])
-        }
-
-        Action::ExplainAnalyzeRequest => {
-            if reject_unsupported_explain(state, services) {
-                return DispatchResult::handled();
-            }
-            let content = state.sql_modal.editor.content().trim().to_string();
-            if content.is_empty() {
-                return DispatchResult::handled();
-            }
-            let Some(dsn) = &state.session.dsn else {
-                return DispatchResult::handled();
-            };
-            let dsn = dsn.clone();
-            if matches!(state.sql_modal.status(), SqlModalStatus::Running) {
-                return DispatchResult::handled();
-            }
-            if is_multi_statement(&content) {
-                show_explain_error_on_plan(
-                    state,
-                    "EXPLAIN ANALYZE does not support multiple statements",
-                );
-                return DispatchResult::handled();
-            }
-            let kind = statement_classifier::classify(&content);
-            let risk = evaluate_sql_risk(&kind, &content);
-
-            let is_dml = !matches!(kind, StatementKind::Select | StatementKind::Transaction);
-
-            if state.session.read_only && is_dml {
-                show_explain_error_on_plan(
-                    state,
-                    "Read-only mode: EXPLAIN ANALYZE is blocked for DML statements.",
-                );
-                return DispatchResult::handled();
-            }
-
-            state.explain.confirm_scroll_offset = 0;
-
-            match risk.confirmation {
-                ConfirmationType::TableNameInput { target } => {
-                    state
-                        .sql_modal
-                        .begin_confirming_analyze_high(content, Some(target));
-                }
-                ConfirmationType::Immediate => {
-                    let Some(explain_query) =
-                        services.sql_dialect.build_explain_analyze_sql(&content)
-                    else {
-                        mark_explain_unavailable(state, services);
-                        return DispatchResult::handled();
-                    };
-                    begin_explain_running(state, now);
-                    return DispatchResult::handled_with(vec![Effect::ExecuteExplain {
-                        dsn,
-                        query: explain_query,
-                        is_analyze: true,
-                        read_only: state.session.read_only,
-                    }]);
-                }
-            }
-
-            DispatchResult::handled()
-        }
-
-        Action::ExplainAnalyzeConfirm => {
-            if reject_unsupported_explain(state, services) {
-                return DispatchResult::handled();
-            }
-            let query = match state.sql_modal.status() {
-                SqlModalStatus::ConfirmingAnalyzeHigh {
-                    query,
-                    input,
-                    target_name,
-                } => target_name
-                    .as_ref()
-                    .is_some_and(|name| input.content() == name)
-                    .then(|| query.clone()),
-                _ => None,
-            };
-            if let Some(query) = query
-                && let Some(dsn) = state.session.dsn.clone()
-            {
-                let Some(explain_query) = services.sql_dialect.build_explain_analyze_sql(&query)
-                else {
-                    mark_explain_unavailable(state, services);
-                    return DispatchResult::handled();
-                };
-                begin_explain_running(state, now);
-                return DispatchResult::handled_with(vec![Effect::ExecuteExplain {
-                    dsn,
-                    query: explain_query,
-                    is_analyze: true,
-                    read_only: state.session.read_only,
-                }]);
-            }
-            DispatchResult::handled()
-        }
-
-        Action::ExplainAnalyzeCancel => {
-            if matches!(
-                state.sql_modal.status(),
-                SqlModalStatus::ConfirmingAnalyzeHigh { .. }
-            ) {
-                state.sql_modal.cancel_confirmation();
-            }
-            DispatchResult::handled()
-        }
-
-        Action::ExplainCompleted {
-            plan_text,
-            is_analyze,
-            execution_time_ms,
-        } => {
-            let query = state.sql_modal.editor.content().to_string();
-            finish_explain_success(
-                state,
-                plan_text.clone(),
-                *is_analyze,
-                *execution_time_ms,
-                &query,
-            );
-            DispatchResult::handled()
-        }
-
-        Action::ExplainFailed(error) => {
-            finish_explain_error(state, error.user_message());
-            DispatchResult::handled()
-        }
-
-        Action::Scroll {
-            target:
-                target @ (ScrollTarget::ExplainConfirm
-                | ScrollTarget::ExplainPlan
-                | ScrollTarget::ExplainCompare),
-            direction,
-            amount: ScrollAmount::Line,
-        } => {
-            let (offset, max) = match target {
-                ScrollTarget::ExplainConfirm => {
-                    // blank + title + blank + separator + blank + warning(2) + blank = 8
-                    const CONFIRM_HEADER_LINES: usize = 8;
-                    let content_lines =
-                        CONFIRM_HEADER_LINES + state.sql_modal.editor.content().lines().count();
-                    let modal_inner = ExplainContext::modal_inner_height(state.ui.terminal_height);
-                    (
-                        &mut state.explain.confirm_scroll_offset,
-                        content_lines.saturating_sub(modal_inner),
-                    )
-                }
-                ScrollTarget::ExplainPlan => {
-                    let modal_inner = ExplainContext::modal_inner_height(state.ui.terminal_height);
-                    let max = state.explain.line_count().saturating_sub(modal_inner);
-                    (&mut state.explain.scroll_offset, max)
-                }
-                ScrollTarget::ExplainCompare => {
-                    let max = state.explain.compare_max_scroll(state.ui.terminal_height);
-                    (&mut state.explain.compare_scroll_offset, max)
-                }
-                _ => unreachable!(),
-            };
-            *offset = direction.clamp_vertical_offset(*offset, max, 1);
-            DispatchResult::handled()
-        }
-
-        Action::CompareEditQuery => {
-            if let Some(ref right) = state.explain.right {
-                let query = right.full_query.clone();
-                state.sql_modal.load_query_for_editing(query);
-            }
-            DispatchResult::handled()
-        }
-
-        Action::SqlModalNextTab => {
-            let tab = services
-                .db_capabilities
-                .next_sql_modal_tab(state.sql_modal.active_tab());
-            state.sql_modal.set_active_tab(tab);
-            DispatchResult::handled()
-        }
-
-        Action::SqlModalPrevTab => {
-            let tab = services
-                .db_capabilities
-                .prev_sql_modal_tab(state.sql_modal.active_tab());
-            state.sql_modal.set_active_tab(tab);
-            DispatchResult::handled()
-        }
-
-        _ => DispatchResult::pass(),
-    }
+    dispatch_explain(state, action, now, services)
 }
 
 #[cfg(test)]
 pub fn reduce_explain(state: &mut AppState, action: &Action, now: Instant) -> DispatchResult {
-    reduce_explain_with_services(state, action, now, &crate::services::AppServices::stub())
+    dispatch_explain(state, action, now, &crate::services::AppServices::stub())
 }
 
 #[cfg(test)]
